RSVP page/global data edits - Robert
RSVP page currently gets all the guests even ones from different events so i need to fix that but at least it displays all the emails. I also need to figure out how the email will send two unique links (Going or Not Going) so that the rsvp page will update going status. Also now with create event/confirm event the event will be added to the list so that if you navigate to events it will update the page. (The event list page is not user specific at the moment because we dont have a user session so i cant add that in yet so if any user logged in they could see any event available