Sunday Update - Robert Piercy
When an event is created it will be updated in the globaldata meaning new events can be viewed (NOT THE DATABASE). ALso i created a session so that when user logs in and checks their events they will only be able to view there events no one elses. Having trouble with RSVP page. I need to loop through the guest lists for each event but couldnt figure it out.